# performance: improve for PathString impl for EvaluationContext