--- a/src/has_headers.rs
+++ b/src/has_headers.rs
@@ -1,12 +1,12 @@
 use hyper::HeaderMap;
 
-use crate::lambda::EvaluationContext;
+use crate::lambda::{EvaluationContext, GraphqlContext};
 
 pub trait HasHeaders {
   fn headers(&self) -> &HeaderMap;
 }
 
-impl HasHeaders for EvaluationContext<'_> {
+impl<'a, Ctx: GraphqlContext<'a>> HasHeaders for EvaluationContext<'a, Ctx> {
   fn headers(&self) -> &HeaderMap {
     self.headers()
   }