--- a/src/lambda/evaluation_context.rs
+++ b/src/lambda/evaluation_context.rs
@@ -65,32 +65,42 @@ impl<'a, Ctx: GraphqlContext<'a>> EvaluationContext<'a, Ctx> {
     self.graphql_ctx.value()
   }
 
-  pub fn args(&self) -> Option<async_graphql::Value> {
-    Some(async_graphql::Value::Object(self.graphql_ctx.args()?.clone()))
+  pub fn arg<T: AsRef<str>>(&self, path: &[T]) -> Option<&'a Value> {
+    let arg = self.graphql_ctx.args()?.get(path[0].as_ref());
+
+    get_path_value(arg?, &path[1..])
   }
 
-  pub fn path_value<T: AsRef<str>>(&self, path: &[T]) -> Option<&'a async_graphql::Value> {
+  pub fn path_value<T: AsRef<str>>(&self, path: &[T]) -> Option<&'a Value> {
     get_path_value(self.graphql_ctx.value()?, path)
   }
 
   pub fn headers(&self) -> &HeaderMap {
     &self.req_ctx.req_headers
   }
-  pub fn get_header_as_value(&self, key: &str) -> Option<async_graphql::Value> {
+
+  pub fn header(&self, key: &str) -> Option<&str> {
     let value = self.headers().get(key)?;
-    Some(async_graphql::Value::String(value.to_str().ok()?.to_string()))
+
+    value.to_str().ok()
+  }
+
+  pub fn var(&self, key: &str) -> Option<&str> {
+    let vars = &self.req_ctx.server.vars;
+
+    vars.get(key).map(|v| v.as_str())
   }
 }
 
-fn get_path_value<'a, T: AsRef<str>>(input: &'a async_graphql::Value, path: &[T]) -> Option<&'a async_graphql::Value> {
+fn get_path_value<'a, T: AsRef<str>>(input: &'a Value, path: &[T]) -> Option<&'a Value> {
   let mut value = Some(input);
   for name in path {
     match value {
-      Some(async_graphql::Value::Object(map)) => {
-        value = map.get(&async_graphql::Name::new(name));
+      Some(Value::Object(map)) => {
+        value = map.get(&Name::new(name));
       }
 
-      Some(async_graphql::Value::List(list)) => {
+      Some(Value::List(list)) => {
         value = list.get(name.as_ref().parse::<usize>().ok()?);
       }
       _ => return None,
@@ -102,6 +112,7 @@ fn get_path_value<'a, T: AsRef<str>>(input: &'a async_graphql::Value, path: &[T]
 
 #[cfg(test)]
 mod tests {
+  use async_graphql::Value;
   use serde_json::json;
 
   use crate::lambda::evaluation_context::get_path_value;
@@ -117,12 +128,12 @@ mod tests {
         }
     });
 
-    let async_value = async_graphql::Value::from_json(json).unwrap();
+    let async_value = Value::from_json(json).unwrap();
 
     let path = vec!["a".to_string(), "b".to_string(), "c".to_string()];
     let result = get_path_value(&async_value, &path);
     assert!(result.is_some());
-    assert_eq!(result.unwrap(), &async_graphql::Value::String("d".to_string()));
+    assert_eq!(result.unwrap(), &Value::String("d".to_string()));
   }
 
   #[test]
@@ -134,7 +145,7 @@ mod tests {
         }
     });
 
-    let async_value = async_graphql::Value::from_json(json).unwrap();
+    let async_value = Value::from_json(json).unwrap();
 
     let path = vec!["a".to_string(), "b".to_string(), "c".to_string()];
     let result = get_path_value(&async_value, &path);
@@ -150,11 +161,11 @@ mod tests {
         }]
     });
 
-    let async_value = async_graphql::Value::from_json(json).unwrap();
+    let async_value = Value::from_json(json).unwrap();
 
     let path = vec!["a".to_string(), "0".to_string(), "b".to_string()];
     let result = get_path_value(&async_value, &path);
     assert!(result.is_some());
-    assert_eq!(result.unwrap(), &async_graphql::Value::String("c".to_string()));
+    assert_eq!(result.unwrap(), &Value::String("c".to_string()));
   }
 }