--- a/src/lambda/lambda.rs
+++ b/src/lambda/lambda.rs
@@ -70,8 +70,7 @@ mod tests {
 
   use crate::endpoint::Endpoint;
   use crate::http::RequestContext;
-  use crate::lambda::evaluation_context::EmptyGraphqlContext;
-  use crate::lambda::{EvaluationContext, Lambda};
+  use crate::lambda::{EmptyResolverContext, EvaluationContext, Lambda};
   use crate::request_template::RequestTemplate;
 
   impl<B> Lambda<B>
@@ -80,7 +79,7 @@ mod tests {
   {
     async fn eval(self) -> Result<B> {
       let req_ctx = RequestContext::default();
-      let ctx = EvaluationContext::new(&req_ctx, &EmptyGraphqlContext);
+      let ctx = EvaluationContext::new(&req_ctx, &EmptyResolverContext);
       let result = self.expression.eval(&ctx).await?;
       let json = serde_json::to_value(result)?;
       Ok(serde_json::from_value(json)?)