--- a/src/lambda/mod.rs
+++ b/src/lambda/mod.rs
@@ -2,6 +2,6 @@ mod evaluation_context;
 mod expression;
 mod lambda;
 
-pub use evaluation_context::EvaluationContext;
+pub use evaluation_context::{EvaluationContext, GraphqlContext};
 pub use expression::Expression;
 pub use lambda::Lambda;