--- a/src/lambda/mod.rs
+++ b/src/lambda/mod.rs
@@ -1,7 +1,9 @@
 mod evaluation_context;
 mod expression;
 mod lambda;
+mod resolver_context_like;
 
-pub use evaluation_context::{EvaluationContext, GraphqlContext};
+pub use evaluation_context::EvaluationContext;
 pub use expression::Expression;
 pub use lambda::Lambda;
+pub use resolver_context_like::{EmptyResolverContext, ResolverContextLike};