**Summary:**  

Implement benchmark for `impl PathString for EvaluationContext` and optimize its code

**Issue Reference(s):**  
Fixes #347 
/claim #347

**Build & Testing:**

- [x] I ran `cargo test` successfully.
- [x] I have run `./lint.sh` to address and fix linting issues.

**Checklist:**

- [ ] I have added relevant unit & integration tests.
- [ ] I have updated the documentation accordingly (if applicable).
- [x] I have performed a self-review of my own code.

## How to review

I suggest to review the pr commit by commit:

1. Make PathString trait generic over type of parameter to be able accept both `&[String]` and `&[str]`
2. Decouple EvaluationContext from `async_graphql::ResolverContext` to be able to create test environment for it
3. Implement benchmark for the impl
4. Optimize and refactor the code of impl

## Results

![image](https://github.com/tailcallhq/tailcall/assets/8974488/a95427ff-eddf-4294-a6e0-11c68f6982b4)

<details>
<summary>All results as text</summary>
<code>
❯ cargo bench --bench impl_path_string_for_evaluation_context -- --baseline init
   Compiling tailcall v0.1.0 (/home/nixos/Projects/open-source/rust/tailcall)
    Finished bench [optimized] target(s) in 33.22s
     Running benches/impl_path_string_for_evaluation_context.rs (target/release/deps/impl_path_string_for_evaluation_context-f4ac70d9689a5945)
Gnuplot not found, using plotters backend
input/value.root        time:   [31.628 ns 31.953 ns 32.409 ns]
                        change: [-32.327% -30.975% -29.546%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe

input/value.nested.existing
                        time:   [65.308 ns 68.170 ns 71.823 ns]
                        change: [-11.594% -9.3241% -5.2798%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 22 outliers among 100 measurements (22.00%)
  16 (16.00%) high mild
  6 (6.00%) high severe

input/value.missing     time:   [25.998 ns 27.113 ns 28.361 ns]
                        change: [-6.3436% -4.0818% -1.5430%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  4 (4.00%) high mild
  8 (8.00%) high severe

input/value.nested.missing
                        time:   [52.774 ns 53.297 ns 53.985 ns]
                        change: [-7.7919% -5.3156% -3.1589%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  3 (3.00%) high mild
  6 (6.00%) high severe

input/args.root         time:   [23.415 ns 23.532 ns 23.648 ns]
                        change: [-83.952% -83.854% -83.757%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

input/args.nested.existing
                        time:   [50.387 ns 51.123 ns 52.093 ns]
                        change: [-71.882% -71.237% -70.655%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high severe

input/args.missing      time:   [13.291 ns 13.386 ns 13.498 ns]
                        change: [-89.743% -89.583% -89.384%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe

input/args.nested.missing
                        time:   [43.307 ns 43.736 ns 44.283 ns]
                        change: [-72.766% -71.938% -71.117%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 18 outliers among 100 measurements (18.00%)
  2 (2.00%) low severe
  2 (2.00%) low mild
  5 (5.00%) high mild
  9 (9.00%) high severe

input/headers.existing  time:   [11.790 ns 11.873 ns 11.967 ns]
                        change: [-11.716% -9.2507% -7.0145%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe

input/headers.missing   time:   [13.003 ns 13.618 ns 14.300 ns]
                        change: [-4.6410% -1.5067% +1.9316%] (p = 0.38 > 0.05)
                        No change in performance detected.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe

input/vars.existing     time:   [5.5251 ns 5.5561 ns 5.5948 ns]
                        change: [-89.906% -89.417% -88.978%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe

input/vars.missing      time:   [6.0952 ns 6.4040 ns 6.7412 ns]
                        change: [-85.395% -84.837% -84.228%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  2 (2.00%) high mild
  4 (4.00%) high severe

</code>
</details>