@meskill 2023-10-07
> Can you share the improvement if any on `value`.

Here is screenshot from the latest run 

![image](https://github.com/tailcallhq/tailcall/assets/8974488/166e0d63-20c7-4347-b07c-79d835689e6c)

Also you can find the whole result under spoiler at the end in the pr description